Refactor AttributeShards to AttributeShardAdder
